remove country code from language store
the locale observable field in language store used to store languages in the format of "languageCode-countryCode", which made the app irresponsive when facing an accepted language code with an unaccepted country code. in this commit we remove the country code and we continue with only language code from now on.