remove progress bar from info step slider
previously, all steps in step slider contained a progress bar. but since Info step is the only step that contains all the questions, the progress bar is removed to show the difference between this step and others. the progress bar will later be added inside the info step page instead.