pass observables to methods for re-rendering widgets
previously, the navigate.push() was responsible to re-render widgets when navigating back to home page which was not optimal and is changed here so the observer will listen for any changes and re-renders the corresponding widget. also when any step's last task's done button was pressed the button would disappear immediately due to change of current step number to the next step.