make the UX of step slider more user friendly
previously, the visible part of continue button on slides other than the one that user was on would lead to an empty task list. this is fixed in this commit, also, now when user clicks on visible parts of other steps next to the step he's on, the step slider slides to the clicked step.