remove the space between the number of tasks and title
Previously, the title, number of tasks and continue button were in the same column, each wrapped in expanded widgets, which caused a big space between the no of tasks and step title. now the two are wrapped in a separate column which itself is also wrapped in an extended widget to ensure the space between the two and the continue button in the bottom of step slider.